% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soft_IDs.R
\name{get_soft_IDs}
\alias{get_soft_IDs}
\title{get the gene symbols of probes for each GPL}
\usage{
get_soft_IDs(gpl)
}
\arguments{
\item{gpl}{the name of GPL, such as GPL570}
}
\value{
a data.frame which contains the gene symbols of probes
}
\description{
get_soft_IDs need the name of GPL, such as GPL570, and it will return a data.frame
}
\examples{
get_soft_IDs('gpl570')
}
